refactor(session): remove context window cache from context manager (#1319)
